Search: remove old pq logic
